+++ NEWS.md
@@ -46,7 +46,7 @@ Standard library changes
 
 #### Miscellaneous
 
-
+* `foldr` and `mapfoldr` now work on any iterator that supports `Iterators.reverse`, not just arrays ([#31781]).
 
 External dependencies
 ---------------------